separate feature computing and voting process; add "centroid-of-voting" to score-using-embedding
